rename: review 관련 api 폴더 정리
